requirements.txt
google-genai>=1.0.0
fastapi>=0.115.0
uvicorn>=0.34.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
python-multipart>=0.0.9
